# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Papel",
    "description": "Papel is a light-weight core banking application that powers banking operations like account creation, customer deposits, and withdrawals. This app is meant to support a single bank, where users can signup and create bank accounts online, but must visit the branch to withdraw or deposit money.",
    "contact": {}
  },
  "host": "https://papel-bank-io.herokuapp.com",
  "securityDefinitions": {},
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "",
        "summary": "Sign up",
        "tags": [
          "Sign"
        ],
        "operationId": "signUp",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "lastname",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstname",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "email"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "password"
          },
          {
            "name": "type",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "user successfuly created"
          },
          "400": {
            "description": "parameter is required"
          },
          "500": {
            "description": "internal Server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "",
        "summary": "user login",
        "tags": [
          "Sign"
        ],
        "operationId": "login",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "email"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "User is successfully logged in"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/accounts" : {
      "post": {
        "description": "",
        "summary": "user create account",
        "tags": [
          "User"
        ],
        "operationId": "createAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "name": "type",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "text"
          },
          {
            "name": "status",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "text"
          }
        ],
        "responses": {
          "200": {
            "description": "account successfuly created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      },
      "get": {
        "description": "",
        "summary": "view all accounts ",
        "tags": [
          "Admin"
        ],
        "operationId": "transaction",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "display all user accounts in the database"
          },
          "400": {
            "description": "bs"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/user/{email}/accounts": {
      "get": {
        "description": "",
        "summary": "User can view all accounts with respect to user's email",
        "tags": [
          "User"
        ],
        "operationId": "userAccounts",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "name" : "email",
            "in" : "path",
            "type" : "string",
            "required":true
          }
        ],
        "responses": {
          "200": {
            "description": "All accounts will be displayed"
          },
          "400": {
            "description": "bad request"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    
    "/accounts/{accountnumber}": {
      "get": {
        "description": "",
        "summary": "user specific account",
        "tags": [
          "User"
        ],
        "operationId": "specificAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "display specified user account"
          },
          "404": {
            "description": "data are not found"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "description": "",
        "summary": "change account's status",
        "tags": [
          "Admin"
        ],
        "operationId": "accountStatus",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "account changed to status"
          },
          "401": {
            "description": "Authentication failed"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "delete": {
        "description": "",
        "summary": "delete account number",
        "tags": [
          "Admin"
        ],
        "operationId": "rejected",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "account successfully deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/accounts/{accountnumber}/transactions": {
      "get": {
        "description": "",
        "summary": "All transactions on one account",
        "tags": [
          "User"
        ],
        "operationId": "transaction",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "View all user account's transactions"
          },
          "404": {
            "description": "Account number not found"
          },
          "401": {
            "description": "Authentication failed"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/transactions/{transactionid}": {
      "get": {
        "description": "",
        "summary": "view specific account's transaction",
        "tags": [
          "User"
        ],
        "operationId": "transaction",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "transactionid",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "display specified transaction"
          },
          "404": {
            "description": "account number nnot found"
          },
          "401": {
            "description": "Authentication failed"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/transactions/{accountnumber}/credit": {
      "post": {
        "description": "",
        "summary": "cashier credit",
        "tags": [
          "Cashier"
        ],
        "operationId": "creditAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "amount",
            "in": "formData",
            "required": true,
            "type": "number",
            "format": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "account successfuly creited"
          },
          "400": {
            "description": "bad reequest"
          },
          "401": {
            "description": "Auth failed"
          }
        }
      }
    },
    "/transactions/{accountnumber}/debit": {
      "post": {
        "description": "",
        "summary": "cashier debit",
        "tags": [
          "Cashier"
        ],
        "operationId": "debitAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "name": "accountnumber",
            "in": "path",
            "required": true,
            "type": "number",
            "format": "number"
          },
          {
            "name": "amount",
            "in": "formData",
            "required": true,
            "type": "number",
            "format": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "account successfuly debited"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Auth failed"
          }
        }
      }
    },
    "/accounts?status=active": {
      "get": {
        "description": "",
        "summary": "view all active accounts ",
        "tags": [
          "Admin"
        ],
        "operationId": "transaction",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "display all user accounts in the database"
          },
          "400": {
            "description": "bs"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/accounts?status=dormant": {
      "get": {
        "description": "",
        "summary": "view all dormant accounts ",
        "tags": [
          "Admin"
        ],
        "operationId": "transaction",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "display all user accounts in the database"
          },
          "400": {
            "description": "bs"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    }
  },
  "definitions": {},
  "tags": [
    {
      "name": "All Endpoints",
      "description": "all endpoints  return objects in JSON format"
    }
  ]
}